+++ src/functions/collections.ts
@@ -3,15 +3,19 @@ import { uniq } from 'lodash';
 import { isDefined } from '../helpers/utils';
 
 export const collections = {
+  any(collection: unknown[] | undefined): boolean {
+    return isDefined(collection) && collection.length > 0;
+  },
+
   distinct(collection: unknown[]): unknown[] {
     return uniq(collection);
   },
 
-  in(value: unknown, collection: unknown[] | undefined): boolean {
-    return isDefined(collection) && collection.includes(value);
-  },
-
   first(collection: unknown[] | undefined): unknown {
     return collection?.[0];
+  },
+
+  in(value: unknown, collection: unknown[] | undefined): boolean {
+    return isDefined(collection) && collection.includes(value);
   }
 };